# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "轻量级智能数据收集平台后端接口。所有接口（除登录外）需在请求头携带 JWT token，格式：Bearer \u003ctoken\u003e。\n\n**状态码说明（提交记录 status）：** 0=待检测 1=正常 2=有异常\n**状态码说明（表单 status）：** 0=草稿 1=已发布 2=已归档",
        "title": "Lite Collector 接口文档",
        "contact": {
            "name": "Lite Collector Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/forms/{formId}/submissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定表单下所有用户的提交记录（不含字段值）。仅表单创建者可访问。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提交记录"
                ],
                "summary": "获取表单的所有提交记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "表单 ID",
                        "name": "formId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.submissionListResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "无权访问该表单",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "表单不存在",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向已发布的表单提交字段数据。每位用户对每个表单只能提交一次。提交后后端自动触发 AI 异常检测（异步），status 会在后台更新。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提交记录"
                ],
                "summary": "提交表单数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "表单 ID",
                        "name": "formId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "字段 key→value 映射，key 对应表单 schema 中的 field_key",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.submissionResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/forms/{formId}/submissions/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户在指定表单中的提交记录，包含所有填写的字段值。每人每表单限提交一次。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提交记录"
                ],
                "summary": "获取我的提交记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "表单 ID",
                        "name": "formId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.submissionWithValuesResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "提交记录不存在",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/forms/{formId}/submissions/{submissionId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定提交记录及其所有字段值。仅表单创建者可访问。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "提交记录"
                ],
                "summary": "获取单条提交记录详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "表单 ID",
                        "name": "formId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "提交记录 ID",
                        "name": "submissionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.submissionWithValuesResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "无权访问该表单",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "表单或提交记录不存在",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{jobId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据任务 ID 查询异步 AI 任务的当前状态及结果。状态值：0=排队中 1=处理中 2=已完成 3=失败。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI任务"
                ],
                "summary": "查询 AI 任务状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI 任务 ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.jobStatusResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "任务不存在",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.errorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "FORM_NOT_FOUND"
                },
                "message": {
                    "type": "string",
                    "example": "form not found"
                }
            }
        },
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.errorDetail"
                }
            }
        },
...
        "handlers.jobStatusResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "input": {
                    "type": "string"
                },
                "job_type": {
                    "type": "string",
                    "example": "detect_anomaly"
                },
                "output": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "handlers.submissionListResponse": {
            "type": "object",
            "properties": {
                "submissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.submissionResponse"
                    }
                }
            }
        },
        "handlers.submissionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 7
                },
                "status": {
                    "type": "integer",
                    "example": 0
                },
                "submitted_at": {
                    "type": "string"
                }
            }
        },
        "handlers.submissionWithValuesResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 7
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "submitted_at": {
                    "type": "string"
                },
                "values": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "登录后获取的 JWT token，格式：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}